Remove the concept of a 'form model'
